@@ -12,7 +12,7 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "^15.0.4",
-    "@angular/cdk": "^15.0.3",
+    "@angular/cdk": "^15.0.4",
     "@angular/common": "^15.0.4",
     "@angular/compiler": "^15.0.4",
     "@angular/core": "^15.0.4",
@@ -23,7 +23,7 @@
     "@angular/router": "^15.0.4",
     "@angular/service-worker": "^15.0.4",
     "@ctrl/ngx-codemirror": "^6.0.0",
-    "@ng-bootstrap/ng-bootstrap": "^14.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^14.0.1",
     "@popperjs/core": "^2.11.6",
     "angular-gridster2": "^15.0.0",
     "angular-split": "^14.1.0",
@@ -46,24 +46,24 @@
     "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.0.4",
+    "@angular-devkit/build-angular": "^15.0.5",
     "@angular-eslint/builder": "^15.1.0",
     "@angular-eslint/eslint-plugin": "^15.1.0",
     "@angular-eslint/eslint-plugin-template": "^15.1.0",
     "@angular-eslint/schematics": "^15.1.0",
     "@angular-eslint/template-parser": "^15.1.0",
-    "@angular/cli": "^15.0.4",
+    "@angular/cli": "^15.0.5",
     "@angular/compiler-cli": "^15.0.4",
     "@angular/language-service": "^15.0.4",
-    "@types/codemirror": "^5.60.5",
+    "@types/codemirror": "^5.60.6",
     "@types/file-saver": "^2.0.5",
     "@types/jasmine": "~3.6.11",
     "@types/jasminewd2": "~2.0.10",
     "@types/markdown-it": "^12.2.3",
     "@types/node": "^18.11.18",
-    "@typescript-eslint/eslint-plugin": "^5.47.1",
-    "@typescript-eslint/parser": "^5.47.1",
-    "eslint": "^8.30.0",
+    "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "@typescript-eslint/parser": "^5.48.1",
+    "eslint": "^8.31.0",
     "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~5.0.2",
     "karma": "~6.3.20",
@@ -73,5 +73,8 @@
     "karma-jasmine-html-reporter": "^1.7.0",
     "ts-node": "^10.9.1",
     "typescript": "~4.8.4"
+  },
+  "resolutions": {
+    "debug": "^4.3.4"
   }
 }
