@@ -43,7 +43,7 @@
     "dockerode": "^3.3.4",
     "express": "^4.18.2",
     "glob-to-regexp": "^0.4.1",
-    "mongoose": "^6.8.2",
+    "mongoose": "^6.8.3",
     "multer": "^1.4.5-lts.1",
     "nats": "^2.10.2",
     "passport": "^0.6.0",
@@ -55,35 +55,36 @@
     "unzipper": "^0.10.11"
   },
   "devDependencies": {
-    "@nestjs/cli": "^9.1.7",
+    "@nestjs/cli": "^9.1.8",
     "@nestjs/schematics": "^9.0.4",
     "@nestjs/testing": "^9.2.1",
     "@types/chownr": "^1.0.0",
     "@types/cron": "^2.0.0",
     "@types/express": "^4.17.15",
     "@types/glob-to-regexp": "^0.4.1",
-    "@types/jest": "^29.2.4",
+    "@types/jest": "^29.2.5",
     "@types/multer": "^1.4.7",
     "@types/node": "^18.11.18",
     "@types/supertest": "^2.0.12",
     "@types/unzipper": "^0.10.5",
-    "@typescript-eslint/eslint-plugin": "^5.47.1",
-    "@typescript-eslint/parser": "^5.47.1",
-    "eslint": "^8.30.0",
-    "eslint-config-prettier": "^8.5.0",
+    "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "@typescript-eslint/parser": "^5.48.1",
+    "eslint": "^8.31.0",
+    "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-prettier": "^4.2.1",
     "jest": "^29.3.1",
-    "prettier": "^2.8.1",
+    "prettier": "^2.8.2",
     "supertest": "^6.3.3",
     "ts-jest": "^29.0.3",
     "ts-loader": "^9.4.2",
     "ts-node": "^10.9.1",
-    "tsconfig-paths": "^4.1.1",
+    "tsconfig-paths": "^4.1.2",
     "typescript": "^4.9.4",
     "webpack": "^5.75.0"
   },
   "resolutions": {
-    "jsonwebtoken": "^9.0.0"
+    "jsonwebtoken": "^9.0.0",
+    "debug": "^4.3.4"
   },
   "jest": {
     "moduleFileExtensions": [
