@@ -1,10 +1,10 @@
 import {Injectable} from '@angular/core';
-import Assignment from '../model/assignment';
+import Assignment, {CreateAssignmentDto} from '../model/assignment';
 import {CreateEvaluationDto, Evaluation} from '../model/evaluation';
 
 @Injectable()
 export class AssignmentContext {
-  assignment: Assignment;
+  assignment: Assignment | CreateAssignmentDto;
   evaluations?: Record<string, Evaluation | CreateEvaluationDto>;
   saveDraft: () => void;
 }
